Reorganize html conversion classes
