+++ app/src/main/java/com/nononsenseapps/feeder/ui/ReaderFragment.java
@@ -39,10 +39,10 @@ import android.widget.TextView;
 import com.nononsenseapps.feeder.R;
 import com.nononsenseapps.feeder.db.FeedItemSQL;
 import com.nononsenseapps.feeder.db.RssContentProvider;
-import com.nononsenseapps.feeder.model.ImageTextLoader;
+import com.nononsenseapps.feeder.ui.text.ImageTextLoader;
 import com.nononsenseapps.feeder.util.TabletUtils;
 import com.nononsenseapps.feeder.views.ObservableScrollView;
-import com.nononsenseapps.text.HtmlConverter;
+import com.nononsenseapps.feeder.ui.text.HtmlConverter;
 
 import org.joda.time.DateTimeZone;
 import org.joda.time.format.DateTimeFormat;
@@ -74,17 +74,11 @@ public class ReaderFragment extends Fragment
     private long _id = -1;
     // All content contained in RssItem
     private FeedItemSQL mRssItem;
-    private TextView mTitleTextView;
     private TextView mBodyTextView;
     private ObservableScrollView mScrollView;
-    private TextView mAuthorTextView;
-    private TextView mFeedTitleTextView;
-
-    private final HtmlConverter htmlformatter;
 
 
     public ReaderFragment() {
-        htmlformatter = new HtmlConverter();
     }
 
     /**
@@ -176,25 +170,22 @@ public class ReaderFragment extends Fragment
 
         mScrollView =
                 (ObservableScrollView) rootView.findViewById(R.id.scroll_view);
-        mTitleTextView = (TextView) rootView.findViewById(R.id.story_title);
+        TextView mTitleTextView = (TextView) rootView.findViewById(R.id.story_title);
         mBodyTextView = (TextView) rootView.findViewById(R.id.story_body);
-        mAuthorTextView = (TextView) rootView.findViewById(R.id.story_author);
-        mFeedTitleTextView = (TextView) rootView.findViewById(R.id
+        TextView mAuthorTextView = (TextView) rootView.findViewById(R.id.story_author);
+        TextView mFeedTitleTextView = (TextView) rootView.findViewById(R.id
                 .story_feedtitle);
 
         if (mRssItem.title == null) {
             mTitleTextView.setText(R.string.nothing_to_display);
         } else {
-            mTitleTextView
-                    .setText(htmlformatter.toSpanned(mRssItem.title, getActivity()));
+            mTitleTextView.setText(HtmlConverter.toSpannedWithNoImages(mRssItem.title, getActivity()));
         }
         if (mRssItem.description == null) {
             mBodyTextView.setText(R.string.nothing_to_display);
         } else {
-            //Log.d("JONAS", "Text is:\n" + mRssItem.description);
             // Set without images as a place holder
-            mBodyTextView.setText(
-                    htmlformatter.toSpanned(mRssItem.description, getActivity()));
+            mBodyTextView.setText(HtmlConverter.toSpannedWithNoImages(mRssItem.description, getActivity()));
         }
 
         if (mRssItem.feedtitle == null) {
@@ -320,7 +311,9 @@ public class ReaderFragment extends Fragment
     @Override
     public void onLoadFinished(final Loader<Spanned> loader,
                                final Spanned data) {
-        mBodyTextView.setText(data);
+        if (data != null) {
+            mBodyTextView.setText(data);
+        }
     }
 
     @Override
