--- app/src/main/java/com/nononsenseapps/text/TableFormatter.java
@@ -1,7 +1,0 @@
-package com.nononsenseapps.text;
-
-/**
- * Created by jonas on 2014-11-15.
- */
-public class TableFormatter {
-}
