+++ app/src/test/java/com/nononsenseapps/text/HtmlToPlainTextConverterTest.java
@@ -1,5 +1,6 @@
 package com.nononsenseapps.text;
 
+import com.nononsenseapps.feeder.ui.text.HtmlToPlainTextConverter;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
